Whole of code changed.
Now podcasts belong to a team, not a user, which allows for collaboration and teamwork.

+++ app/Http/Livewire/Podcast/Edit.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire\Podcast;
 
+use Illuminate\Support\Facades\Log;
 use Livewire\Component;
 use Illuminate\Support\Facades\Storage;
 use Livewire\WithFileUploads;
@@ -80,23 +81,30 @@ class Edit extends Component
     // Delete Podcast
     public function deletePodcast()
     {
-        foreach ($this->podcast->episodes as $episode) {
-            // Delete episode files
-            Storage::disk('s3')->delete($episode->file_name);
+        if (auth()->user()->hasTeamRole( auth()->user()->currentTeam, 'admin' )) {
 
-            // Delete episodes from DB
-            $episode->delete();
+            foreach ($this->podcast->episodes as $episode) {
+                // Delete episode files
+                Storage::disk('s3')->delete($episode->file_name);
+
+                // Delete episodes from DB
+                $episode->delete();
+            }
+
+            // Delete podcast thumbnail image
+            Storage::disk('s3')->delete($this->podcast->thumbnail);
+
+            // Delete podcast
+            $this->podcast->delete();
+
+            session()->flash('success', 'The podcast ' . $this->podcast->name . ', and all its content has been successfully deleted from our platform.');
+
+            return redirect(route('podcasts'));
+
+        } else {
+            Log::error("A user tried to delete a podcast - User: " . auth()->user()->name . " - Podcast: " . $this->podcast->name);
+            abort(503, 'Unauthorized action. This will be reported!');
         }
-
-        // Delete podcast thumbnail image
-        Storage::disk('s3')->delete($this->podcast->thumbnail);
-
-        // Delete podcast
-        $this->podcast->delete();
-
-        session()->flash('success', 'The podcast ' . $this->podcast->name . ', and all its content has been successfully deleted from our platform.');
-
-        return redirect(route('podcasts'));
     }
 
     public function render()
